==> notebooks/01-scrape-gov-web.ipynb <==
from bs4 import BeautifulSoup
from selenium import webdriver
from time import sleep
from time import time
from random import randint
from IPython.core.display import clear_output
import pandas as pd
import numpy as np
import psycopg2
import requests
import os
import re
import ast

headers = {'User-Agent': 'Mozilla/5.0',
           'From': 'itchyandscratchy@gmail.com'
          }

data_path = '/Users/amanda/Documents/Projects/insight/data/'

url = 'https://noc.esdc.gc.ca/Structure/Hierarchy/1a06c040a8bd4022ac498a80eb089bc1?objectid=%2Fd0IGA6qD8JPRfoj5UCjpg%3D%3D'

joblinks = []
# Main summary page
with requests.get(url, timeout=5, headers=headers) as u:     

    soup = BeautifulSoup(u.content)
    for link in soup.find_all('a'):
        joblinks.append(link.get('href'))
            
linkends = [s for s in joblinks if 'NocProfile' in s]            

print(linkends[0])

base_url = 'https://noc.esdc.gc.ca'

overview_df = pd.DataFrame({'noc_link':[],
                            'noc':[],
                            'job_group':[],
                            'description':[],
                            'alt_titles':[]})

duties_df = pd.DataFrame({'noc':[],
                          'duties':[],
                          'requirements':[]})

for link in linkends:
    
    try:

        # Main summary page
        url = base_url + link
        with requests.get(url, timeout=8, headers=headers) as u:     
            soup = BeautifulSoup(u.content) 
            #print(soup.prettify())

            # Main website content
            a = soup.find('main', attrs={'class':'container'})

            # Title and overview section
            title = a.find('h2',attrs={'style':'margin-top:25px;margin-bottom:0;'})
            intro = a.find('p', attrs={'class':'mrgn-tp-md'})

            if title is not None and intro is not None:
                noc_summary = title.text.split('–') #Note this is an en-dash, not a hyphen
                noc_text = noc_summary[0]
                noc_code = 'NOC ' + noc_text.lstrip('0').strip()
                job_group = noc_summary[-1].strip()
                print(noc_code + ': ' + job_group)
                job_intro = intro.text.strip()
            else:
                continue

            title_sec = a.find('div',attrs={'id':'IndexTitles'})
            title_list = title_sec.findAll('li')
            alt_titles = [s.text.strip() for s in title_list]

            overview_df = overview_df.append({'noc_link':link,
                                              'noc':noc_code,
                                              'job_group':job_group,
                                              'description':job_intro,
                                              'alt_titles':alt_titles}, ignore_index=True)

            # Duties and requirements
            job_duties = []
            duty_sec = soup.find('h4', text = 'Main duties').findNext('div', attrs={'class':'panel-body'})
            duty_list = duty_sec.findAll('li')
            if duty_list:
                job_duties = [s.text.strip() for s in duty_list]
            else:
                job_duties = 'NaN'

            job_req = []
            req_head = soup.find('h4', text = 'Employment requirements').findNext('ul')
            req_list = req_head.findAll('li')
            if req_list:
                job_reqs = [s.text.strip() for s in req_list]
            else:
                job_reqs = 'NaN'

            duties_df = duties_df.append({'noc':noc_code,
                                          'duties':job_duties,
                                          'requirements':job_reqs}, ignore_index=True)            
            
    except:
        continue
    
 

# Save cleaned dataframes to file 

overview_df.to_csv(os.path.join(data_path,'processed','noc-overview.csv'), index=False)


# Try urls to find out if they exist
# Note this is a long process and only generally needs to be run once

url_list = []

for i in range(1,30000):  
    try:
        url = 'https://www.jobbank.gc.ca/marketreport/summary-occupation/{}/ca'.format(i)  
        request = requests.get(url, headers=headers, timeout=5, allow_redirects=False)
        if request.status_code == 200:
            url_list.append(url)
            print(url)
        else:
            print('SKIP')
    except:
        continue

clean_job_urls = [x.replace('https://www.jobbank.gc.ca/marketreport/summary-occupation','') for x in url_list]

# Save url_list to file so I can load it later

file_path = (os.path.join(data_path,'processed','job_url_list.txt'))

with open(file_path, 'w') as filehandle:
    for listitem in clean_job_urls:
        filehandle.write('%s\n' % listitem)
