==> notebooks/04-plotting-job-transitions.ipynb <==
import os, re, ast
import psycopg2
import requests
import json, urllib
import pandas as pd
import numpy as np
import plotly
import plotly.graph_objects as go

def convert_to_list(df,columns):

    df.fillna('', inplace=True)

    for col in columns:
        if isinstance(df[col][0], str):
            df[col] = [ast.literal_eval(s) for s in df[col]]

    return df

# Load education and occupation details

jobs_df = pd.read_csv('/Users/amanda/Documents/Projects/insight/data/processed/jobbank-fields.csv')
details_df = pd.read_csv('/Users/amanda/Documents/Projects/insight/data/processed/jobbank-details.csv')
job_name_df = pd.read_csv('/Users/amanda/Documents/Projects/insight/data/processed/education-to-job.csv')


# Clean and organize skills information

jobs_df['description'].str.split('found under ')
edu_categories = [s.split('found under ')[1] for s in jobs_df['description']]
edu_categories = [s.strip('.') for s in edu_categories]
jobs_df['education_groups'] = edu_categories

jobs_df = convert_to_list(jobs_df, ['top_jobs','top_job_links','job_percent'])

education_df = jobs_df[['degree','education_groups','top_jobs','job_percent']].drop_duplicates(subset =['education_groups','degree']) 

details_df = convert_to_list(details_df, ['requirements','skills'])


# Build Sankey diagram

import matplotlib.colors as mc
import colorsys

def adjust_lightness(color, amount=1.28):
    c = color
    c = colorsys.rgb_to_hls(*mc.to_rgb(c))
    rgb = colorsys.hls_to_rgb(c[0], max(0, min(1, amount * c[1])), c[2])
    return rgb

import numpy as np
import matplotlib.cm
import seaborn as sns
from  more_itertools import unique_everseen

def get_cmap_string(palette, domain):

    domain_unique = list(unique_everseen(domain ))
    rgbmap = sns.color_palette(palette, len(domain_unique))
    color_dict = res = dict(zip(domain_unique, rgbmap))   
    color_list = [color_dict.get(i) for i in domain]

    return color_list

def build_sankey(df, degree):

    data = df.where(df['education_groups']==degree).dropna()
    
    data['identifier'] = data['identifier'].str.replace(" : ","- ")

    all_nodes = data['identifier'].values.tolist() + data['top_jobs'].values.tolist()
    source_indices = [all_nodes.index(identifier) for identifier in data['identifier']]
    target_indices = [all_nodes.index(top_job) for top_job in data['top_jobs']]
    
    cmap = get_cmap_string(palette='plasma_r', domain=all_nodes)
    #cmap2 = list(map(adjust_lightness, cmap))

    fig = go.Figure(data=[go.Sankey
                          (node = dict(pad = 15,
                                       thickness = 20,
                                       line = dict(color = "black",
                                                   width = 0),
                                       label = all_nodes,
                                       color = ['rgb' + str(s).strip('[]') for s in cmap]
                                       ),
                           link = dict(source = source_indices,
                                       target = target_indices,
                                       value = data['job_percent']*100,
                                       #color = ['rgb' + str(s).strip('[]') for s in cmap2]
                                       ),
                          )
                         ]
                   )

    filename = 'static/{}.png'.format(degree)

    fig.update_layout(
    autosize=False,
    width=800,
    height=800)

    return fig

fig = build_sankey(job_name_df,'Accounting')

fig.show()

job_name_df.where(job_name_df['education_groups']=='Accounting').dropna()

job_name_df.to_csv('/Users/amanda/Documents/Projects/insight/data/processed/education-to-job.csv', index=False)
