==> notebooks/07-skills-analysis.ipynb <==
import pandas as pd
import numpy as np
import os, re, ast
from fuzzywuzzy import process

def convert_to_list(df,columns):

    df.fillna('', inplace=True)
    
    for col in columns:
        if isinstance(df[col][0], str):
            df[col] = [ast.literal_eval(s) for s in df[col]]
        
    return df

def find_edu_skills(job_df,skills_df,education):

    education_jobs = job_df.loc[job_df['education_groups'] == education]
    merged_df = pd.merge(left=education_jobs, 
                         right=skills_df, 
                         how='left', 
                         left_on='noc', 
                         right_on='noc')
    
    merged_df_unique = merged_df.drop_duplicates(subset=['link'])
    
    skills_expanded = merged_df_unique[['noc',
                                        'edu_level',
                                        'education_groups',
                                        'top_jobs',
                                        'job_percent',
                                        'skills']].explode('skills')
    
    skills_dict = skills_expanded['skills'].value_counts().to_dict()
    
    # Remove skills from the list where the count is less than 2 as they're probably not strongly associated with the educaiton
    skill_values = {key:val for key, val in skills_dict.items() if val > 2}
    
    return skill_values

def find_job_skills(job_lookup,skills_df,job):
    
    job_link = job_lookup.loc[job_lookup['job_group'] == job.lower()]
    
    merged_df = pd.merge(left=job_link, 
                         right=skills_df, 
                         how='left', 
                         left_on='noc', 
                         right_on='noc')
    
    
    merged_df_unique = merged_df.drop_duplicates(subset=['noc'])
    
    skills_expanded = merged_df_unique[['noc',
                                        'job_group',
                                        'skills']].explode('skills')
    
    job_skill_values = skills_expanded['skills'].value_counts().to_dict()
    
    return job_skill_values

data_path = '/Users/amanda/Documents/Projects/insight/data'

# Load education and occupation details

overview = pd.read_csv(os.path.join(data_path,'processed','noc-overview.csv'))
description = pd.read_csv(os.path.join(data_path,'processed','job-description.csv'))
regulation = pd.read_csv(os.path.join(data_path,'processed','job-regulation.csv'))
skills = pd.read_csv(os.path.join(data_path,'processed','job-skills.csv'))

job_name_df = pd.read_csv(os.path.join(data_path,'processed','education-to-job.csv'))


# Convert dataframe columns back to lists where necessary
overview_df = overview
description_df = convert_to_list(description,['duties','titles'])
regulation_df = regulation
skills_df = convert_to_list(skills,['expertise','skills','knowledge'])

job_lookup = overview_df[['job_group','noc']].drop_duplicates()
job_lookup['job_group'] = job_lookup['job_group'].str.lower().str.strip()

[print(x) for x in skills_df.noc]

job = 'Statistical officers and related research support occupations'
job_link = job_lookup.loc[job_lookup['job_group'] == job.lower().strip()]

job_link

# Get skills associated with the given education and compare to new job

degree = 'Civil engineering and related studies'
alt_job = 'Statistical officers and related research support occupations'

edu_skill_dict = find_edu_skills(job_name_df,skills_df,degree)
job_skill_dict = find_job_skills(job_lookup,skills_df,alt_job)

edu_skill_sort = sorted(list(edu_skill_dict.keys()))
job_skill_sort = sorted(list(job_skill_dict.keys()))

matching_skills = [key for key in edu_skill_dict.keys() & job_skill_dict.keys()]
new_skills = set(job_skill_dict.keys()) - set(edu_skill_dict.keys())

new_skills

job_skill_sort

def get_description(job_lookup,df,job):

    job_link = job_lookup['noc'].loc[job_lookup['job_group'] == job.lower()]
    
    print(job_link)

    merged_df = pd.merge(left=job_link, 
                         right=df, 
                         how='left', 
                         left_on='noc', 
                         right_on='noc')
    
    
    description = merged_df['description'][0]
    
    return description

get_description(job_lookup,description_df,'Administrative officers')

job_name_df['education_groups']

q = 'soc'

job_names = list(job_name_df['education_groups'].unique())

print(type(job_names))
query = [s for s in job_names if q.lower() in s.lower()]
query

def find_closest_match(usr_input,name_list):

    highest = process.extract(usr_input,name_list)[0:3]
    top3matches = [x[0] for x in highest]

    return top3matches

edu_names = list(job_name_df['education_groups'].unique())
close_match = find_closest_match('enginering',edu_names)
close_match
